rename livestack ws
